fix(main/module): don't signal a section switch if it's the same section
Fixes #10428

The issue above happened when you double click on a chat section. It causes the fetch to happen twice, which is wasteful, but also causes the model to get reset while it's being processed, thus creating the crash.

The fix here is to not signal that we are changing section if we are trying to switch to the section which is already active (makes sense ofc)