fix(SQ/StatusBaseButton): Incorrect text behaviour when width is specified
Text behaviour was incorrect when width was specified from outside, ie button used inside a layout and filled in width, the text was moved to the left.

Now, an intermediate item as been added and the button row content will be centered on it and  implicit width and height are set depending on the button content size.

Fixes #10480