feat: introduce and handle `checkPermissionsToJoinCommunity()`
This is an improved version to check wether a user has permission to
join a community and updating the join community view accordingly.

We now asynchronously do all the checks in status-go and process
a single result upon token permission updates, additions and deletions.

Depends on: https://github.com/status-im/status-go/pull/3494

Closes #10481 #4939