fix(contacts): fetches info for contacts missing it + loading in AC
Fixes #10400

This calls `requestContactInfo` for all contacts for which we do not have the display name.
While it does that, it sets `fetching`, a new client-side property, to let the UI know that this contact's info are being fetched.

That `fetching` property is used in the AC to show a loading state while it's fetching.

Once the fetching resolves, it removes the loaders and the contact info is updated.

Sadly, in some cases, the fetching can "fail", ie it might not find the info, so we just default back to the 3 word name. The reason for that is that those contacts have not been online for the last 7-9 days.