Retrieve community token contract addresses from token metadata
Prior to this change, the control node would try to fetch the community
token data from its database entries.

However, when a community has been recovered from importing an account
via seedphrase, the database for community tokens will be empty.

So it's better to retrieve the contract addresses from the community
description's metadata (which should always be there).

This change determines the community token from the community metadata.