fix: crash when request to join fails + add error toast when it fails
Fixes #11514

The problem was that we were not handling the error from request to join correctly.

Then I added an event that sends a signal to the QML to show a toast about an error. I don't show the actual error to the user because usually it's not super helpful. It will be in the logs.

The status-go change makes it so that we don't save the request to join if the permission check failed.