refactor: rename SIGNAL_COMMUNITY_TOKEN_DEPLOYED
We've been using this signal whenever a user has sent a transaction that
attemps to deploy a smart contract onchain.

However, at this point it's not yet known if the transaction will
actually go through. So to reflect what's actually happening, this
commits renames a few things:

- `SIGNAL_COMMUNITY_TOKEN_DEPLOYED` becomes
  `SIGNAL_COMMUNITY_TOKEN_DEPLOYMENT_STARTED`
- `CommunityTokenDeployedArgs` becomes `CommunityTokenDeploymentArgs`
- `onCommunityTokenDeployed` becomes `onCommunityTokenDeploymentStarted`

This should remove any confusion around whether or not deployment is
indeed finished.