sign-macos-pkg.sh: move gatekeeper verification
This now fails when the app is signed with a new certificate
create from our new Apple organization but is not notarized:
```
tmp/macos/dist/Status.app: rejected
source=Unnotarized Developer ID
```
I actually have absolutely no idea why this verification worked
with the old certificate, but it did.

For that reason I'm moving it to after notarization.

Signed-off-by: Jakub Sokołowski <jakub@status.im>