performance(NestedListViews): Adding the possibility to position multiple listviews in a single scrollable area without unrolling them
This commit adds the basic infrastructure for a lazy scroll area that will not unroll nested listviews.

How does it work:

There are two public components to be used. The `StatusLazyScrollView` and the `NestedListViewAnchor`. The  `StatusLazyScrollView` component is the main scrollable area. It implements a custom Flickable as `contentItem` and a custom vertical ScrollBar. The `NestedListViewAnchor` needs to be defined for each nested ListView created as a child of `StatusLazyScrollView`. It defines the target ListView to be managed and the Flickable where the ListView is anchored. The ListView implicitHeight, interactive and boundsBehaviour properties are managed by the anchors so that it can properly set the size and scroll the ListView when needed.

The main idea is to have a scrollable area that can switch the scroll event handler from the main Flickable to each child ListView when needed.

E.g. The main Flickable has two ListViews. One at the top, one at the bottom. When scrolling from beginning and the first listview is in viewport, the scrolling handling will be switched to the Listview. When continuing to scroll, the ListView is scrolled until the end. At this point, the scrolling handling is moved back to the main Flickable until the next Listview is reached.

In order to achieve this, other helper components are used:
`ScrollBreakpoint` - This component can watch a target flickable for scrolling related events (`onContentYChanged`, `onOriginYChanged`, `onContentHeightchanged`) and emit a hit signal when the breakpoint conditions is met. This is the base component used to detect when the scrolling handling needs to be switched from one Flickable to another
`NestedListViewScrollBreakpoint` - This component is configuring a few `ScrollBreakpoints` on both main Flickable the the nested ListView. There are two main types of scroll breakpoints. The first one is the external breakpoint. It's meant to detect when the nested Listview is in viewport and send an event. The internal breakpoints will detect when the ListView is scrolled at beginning or at the end.

Limitations:
1. The main Flickable needs to be in the nested ListView parent hierarchy. Only 3 levels are currently supported. The number of levels can be extended or totally eliminated
2. The `StatusLazyScrollView` height cannot be managed by a Layout. Not sure it's an implementation bug or a Qt bug, but if the `StatusLazyScrollView` is managed by a Layout the initial height is set to maximum and all the nested listviews are unrolled until all the heights are properly calculated. Then the ListViews are properly updated.
3. The interactive scrolling using the ScrollBar is not fully supported yet. It needs extra work to overcome the limitations introduced by the unreliable `contentHeight` property of the ListView