fix(chat_section): remove duplicated event handling
Fixes #12719

The problem was that we had two event handlers that did the same thing and if the main module's event processed first, the second event would basically reset the active item, causing the UI to try to fetch an empty chat content.

If you comment out the main module event handler, it doesn't work correctly, because it's the one doing the actual section switching.

Removing the chat section event handling is ok, because the main module already calls its child's (chat section) same functions.