chore: replace waitFor with waitForObjectExists
There is a problem with is_visible method now, it is not raising any assertions in case it did not find any object, therefore i suspect it as a potential endless loop (always returns True it seems). I will try to fix that, however it is being used across all the framework so it takes time.

The main idea here is to get rid of endless loop and 15 minutes if waiting for nothing in this certain test