fix: Multiple permission update toasts are spawned
The permissions updates are done twice. Once when the permission update is called to the backend and the second time when the event for the new community description is received.
The permissions are not updated anymore when the backend rpc is called. The update happens only when the community description signal is received