fix(async_tasks): add try catch to getChannelGroups async task
Kind of fixes #13640

The issue itself was no reproducible even after dozens of restarts and following of the repro steps.
It seems like this was an unfortunate DB close event that caused the app to close.

Adding a try catch won't actually fix that, since the DB is closed and also the `communtiies`already had a try/catch, but it will help for later.