fix(build) fix python package error on make run on MacOS
Update packages all the time ensuring that a change in package.json will
be reflected without manual intervention.

The `fileicon` v0.3.0 tool we used on MacOS to embed icons in our binary
is failing. Upgrading to latest v0.3.3 fixes this. It will require
a `make update` though to fix local builds.