fix(import): fix import community popup
Fixes #15526

Fixes the import popup showing the wrong state. That was because a param was removed from the function, but not from the call to that function, so `true` was passed in `invitationPending` when it shouldn't have.

I also added a call to `spectateCommunity` before the community intro popup so that the community spectated beforee sending the request