chore: sleep to avoid race conditions on intel (#15194)
With the help of build flags this commit we sleep for 0.7 seconds before we make calls to the following status-go functions that crash on intel MacOS.
- `keycard_go.keycardInitFlow`
- `keycard_go.keycardCancelFlow`
- `status_go.initializeApplication`

As per my investigations this crash happens due to a race condition and was not consistently reproducible. Sometimes it would happen and other times it would not.
I know for sure it used to happen in these places and adding a tiny sleep before calling them made sure that the crashes weren't reproducible for me.

I think a 2.1 second delay during onboarding would be acceptable vs the existing condition of the desktop app straight up crashing.