fix(perf): improve chat switching speed by delaying member list loading
Fixes #16132

The problem ended up being the member list (for communities). It takes a long time to load when you have hundreds of members, especially since it has multiple filters on the SortFilterProxy.

The solution I went with was to reuse the same method we have for message list, which is to keep 5 in memory. So we have multiple instances of the ListView and 5 max active at the same time. This uses the same model and property as the MessageView, so there is no code duplication.
This makes sure that we don't have to reload and parse the member list each time we switch. So switching between the last 5 channels is instant like before.

The second solution is to had a fake delay on first load. I tried only using `asynchronous` on the Loader, but it didn't really give an improvement.
By adding a Timer with a delay, we can feel the speed up. It switches rapidly to the channel and we can see the message loading skeleton (if needed).

The only missing part was to add a loader to fill the empty gap when delaying, so I reused the MessagesLoadingView and adapted it so it fits the MemberList format.

Later, we can even load the member list async in the backend and use that Loading animation and not rely on the timer.