perf(chat_section): improve onChatsLoaded performance (#16186)
Fixes #16181

This commit improves the time taken by onChatsLoaded that is called on login.
The culprit is `buildChatSectionUI` in the chat_section module.
There is no silver bullet here. It's slow because the community is big and has a lot of channels to load and generate.

This commit helps by removing some old code that was inefficient (calculating if a channel has a permission instead of just using the `tokenGated` property for example).
It also adds all the section items in one go instead of one another.

There are some other small improvements, but again, no way to make it way faster. Thankfully, that time is spent with the loading spinner at the same time.